BACKEND/requirements.txt
 fastapi
 uvicorn
 psycopg2 
 python-multipart
 sqlalchemy
 python-jose
 passlib[bcrypt]
